Fixed scene node object exporting child-of-owned render node which was putting the child's child's data in the wrong place in the scene node object's data.
Changed the scheme editor so that it now catches new graph node creation (eg. instanced from a scheme) and treats it as a newly-instanced node when dragging from the icons. They're positioned like normally instanced nodes and have a widget each (they had a widget each before which is why the eyeball could be dragged once it was dropped in)
The "new scene node" ref has now been ditched and instead uses the selected node list.

I've added a duplication of widget creation check too which I suspect was the cause of the objects that were moved twice as far as they should have, but I never tracked that down so not sure.

If we have child scene nodes in a scheme with their own offset translations the set of the transform on the mouse move can be changed to be ALL relative changes.