Added TLAsset::LoadAllAssets( TypeRef ) which goes through the file system, finds asset files, finds asset files which will produce an asset of type TypeRef and then loads those assets.
This only works for existing Asset files, so if you have a mesh that needs converting from an SVG to a mesh, then it will miss it (or at least, will only recognise it if it has been compiled into a .asset already). The system CAN be expanded to include these though without too much trouble if required. (would be irrelevant on releases)

This is good to find out what assets exist (via GetAllAssets) but will mean assets which haven't already been loaded/used won't get missed/overwritten. So to find say, all levels, call LoadAllAssets("Level") and then use GetAllAssets<TLAsset::TLevel>() to find all levels (loaded or not).
In future, if/when it needs to be implemented, we will just load the TFileAsset header, as that's all we need to check the asset type being produced.

Surpressed a filesystem timestamp break and will look further into that at a later date.