Changed the scheme editor class to produce a menu asset for the icons. Icons can now be used directly as a menu, or just using the menu item data to create icons from.
Removed the scrollbar code. (Sorry duane, but it doesn't belong in the base code)

Split up the code into seperate modes to make the code a bit more robust and aid the menu-representation version in rocket.

When we drag/select a node (by whatever means, including when dragging a new node from the icon menu) we now switch into node-mode.

Icon widgets now seperated from the gui widgets to make enabling/disabling them easier.

The CreateIconRenderNodes function now isn't called automatically, it's up to the overloaded scheme editor to handle this. I will at some point revert to a default system which uses the text menu wrapper.

Changed the quit-editor command from Close to Quit as this was a bit of a conflict when working with menus.