Finally fixed the issue with the window on PC not being created at the correct size on my machine.  It turns out there are two methods of getting windows size messages at startup when a window is created that is outside of the application - GetStartupInfo() and GetWindowPlacement().  The former will potentially contain window size, position etc if the process that starts the application provides it - in my case this wasn't the cause but I have added the required code to disable the incoming size anyway.  The latter was the offender and is essentially the windows restore state information for the window handle.  This is now reset at startup.
FYI essentially the size, pos etc would be used when the first call to ShowWindow() was done, regardless of what size and pos we had set in code, they would effectively be overwritten because the message to resize with the startup data would come AFTER the window had already been setup via our code.  I'm not sure why you never suffered this Graham but it may be that you never resized your windows and therefore never noticed the issue.  If you resize now it will reset when you restart the app.  We can remove the resizing too if needed but at some point it would be better to handle the size change correctly, alterng aspect ratio etc, so for now I have left this as-is.  This took quite a while to track down as the MSDN info on this topic is very vague.

Big re-ordering of header file includes across the project to try and cut down the build times. I have managed to halve most of the library and project times on my machine so you should definitely notice a difference.  This has had a knock on effect of a few other changes as detailed below.

Added a TLCoreMisc.cpp and .h file to contain routines that are core to the system but not dependant on the platform core files which will include things like windows.h (very slow).  This has allowed for some header dependencies to be removed and considerably improved compile times where files only needed the core TRef defines for example.

Removed the dependency of the TLMemory on windows.h.  This was a major bottleneck where almost every file in the project includes TLMemory via the TArray files for example.  This has had the biggest gain in build performance.

Renamed the filesys CreateFile() call to CreateNewFile() to avoid being ambiguous with the Windows macro for CreateFile.  When doing a build these would either appear as an error at compile time or would break the asset generation at runtime as the CreateFile would effectively be removed from the build depending on when the <windows.h> file was included during the build phase.  Although the reorder of the headers to improve builds times highlighted this issue, its possible that at some point down the line we may have altered the headers and experienced the same problem without it being obvious as to the cause.

Renamed the TPtr GetObject() call to GetObjectPointer() to avoid being ambiguous with the Windows macro of the same name.  Due to header changes this caused compile time errors only.

Added a TScreenShape.h file to the TootleRender library that contains the screenshape enum and routines.  This allows us to remove some of the header dependencies on TLRender and TRenderTarget where the files simply needed the shape enum, but would consiquently build with all of the opengl files and windows files which weren't necessary.  There are more headers here that can be updated and moved around to improve build times.

Removed some of the gameside dependencies on TLRender and TLPhysics which will include a considerable amount of other header files that are generally not needed outside of their respective libraries.

Fixed a big issue with the way I had setup the memory system singleton which in fact wasn't a singleton at all - by breakpointing the constructor and destructor there would be many instances of the static variable due to it being declared in the header and compiled into many source files.  This is now correctly setup as a singleton class and can now detect when the system is trying to be used after it has been destructed which highlights the only major drawback - there is no guarantee of the order of static creation so the memory system can never be guaranteed to be the first thing created.  To alleviate this I have added a TootleMemory library which the core library will be dependant on and a boot.cpp which is compiled first to try and ensure it is created first.  

NOTE: Digging around on the net this is apparently the best way to get the static to be the first thing created.  Whether this works or not I am not totally sure yet, so where possible I would suggest ensuring that any global variables are properly sorted out during an appropriate managers shutdown or destruction to avoid relying on the memory system being in existing when a static is destroyed.  

Additionally, I've been through and fixed any outstanding issues with global objects not being deleted/emptied at a suitable time and therefore the games should generally now close correctly when closing the window on the PC.