Added some minor tweaks to the scheme editor and added some overloadable functionality.
Added IsChangingMode() to the state machine. I had a case where during a mode's OnBegin, through various funcs, the state machine class queried the current mode ref, which was actually the old mode, and I didn't want it to apply for the new mode... so it just ignores it if it's changing mode. Shouldn't need to use it often, but will catch that state between-states.