Fixed import/export bug of TMenu asset where the menu item string, when imported, would import/export as dumb data and not as a string/array. Caused vtables to get corrupted so when we actually came to use the Menu asset item we'd get wierd crashes. (Jump to strange location). Menu assets [with strings, see below] need outputting!
A lot of menu's were using <MenuItem Text="xyz" when it should have been String="abc so strings for menu items weren't being imported at all. Now throws up a Debug_Break but is backwards compatible. The reason it should be string is because it's a data-string. Text could (or will) refer to text ref.

Implemented the TMenuWrapperText menu wrapper. Now if you simply want to render a functional menu with text and not create and maintain a scheme etc just use this type when creating your menu wrapper instead of using the scheme one.
Really only designed for debug/pre-release but could probably use it for final functionality too.