Rearranged some of the physics headers to further alleviate dependencies and build times - this will mainly affect the physics library itself and the projects that use the physics library files.
To support this I have added some new .cpp and .h files for classes or specifically for inline routines of classes that can be included as required.

Changed the physics graph b2world pointer to be a raw pointer rather than a TPtr.  Using the TPtr meant the b2world had to be fully defined for the template whereas using the raw pointer meant I could simply forward declare the class and avoid including quite a lot of the box2d header files.