Added RemoveChildren() to TBinaryTree to remove all children with a certain ref.
Added/implemented a ChangeMode() function to the scheme editor so mode changes are stalled until next update rather than occurring during messages which would often cause issues (menus, widgets being destroyed, recreated on the same frame)
Fiddled around with various Select/Unselect behaviour again

Added an IsIntersecting() function to a widget so you can pass in your own fake click to see if it will intersect with the node/datum of a widget. Useful for re-using an already setup widget to detect movement/clicks on objects.

Changed the drag widget to NOT send out a click up when the drag had never started (ie. it was sending an UP message without first having a DOWN message.) Maybe need to move this functionality into the base widget code.

Removed the post-rotation bounds check for TLRender::TOrthoCamera::GetWorldPos so you will still get a world position even if the new rotated coords are outside of the original square render target box. This may not be wanted at all, but can produce game coordinates that aren't in the world's render target box. Bad for UI's, good for games.