Added reverse order message sending to the TPublisher class.  This is now used by the shutdown handling to shutdown the managers in reverse order of subscription which corrects an issue with the save state of the game not being saved correctly.
At the moment the managers are still shut down in parallel so there is no further dependency checking between managers but that can be added at a later stage if required.

Reverted the change to the file system g_pFileFactory being nulled during the destuctor which is now done in the shutdown as before.