Fixed sphere accumulation code. Would previously generate spheres that were a bit too small.
Added a check in the bounds calculation of a render node where it was doing unnecessary re-calculation of the local bounds (mesh and children)

Added code to the mesh and SVG's to calculate projected UV coordinates. Far top-left vertex of an svg is [s,t]=[0,0] and furthest bottom right is [1,1]. This is going to produce skewed results for non-square geometry but we can tweak this as required.
Enable by adding TootleProjectUVs="True" to your <svg> xml tag.

Also changed TootleOrthoScale to use the document's width instead of the arbitrary 100 value. This means you should set the document with to the ortho range for your desired render target.

Added a ReplaceData() function to TBinaryTree's. This just removes old data, and adds new data (like we use in many places). We can optimise this in future as we see fit, and current usage (like UpdateNodeData code) should be replaced with this function now for future benefits.

Added a check to MoveNode in the graph so we can't try and move a node to it's existing parent.

Added support in TTransform for "ztranslate" data which will just modify the Z of a translate. Handy for 2D stuff!

Added some additoinal swizzle accessor to Type4 types.

changed TWidgets back to putting ALL the node's data in the widget data so we can re-use nodes, datums etc in outgoing messages.

Modified the PC mouse code to consolidate all the mouse moves in the input buffer into one movement. Although there are cavaets with it (see comments in the code) it seems to work fine and reduces around 4-24 mouse move sensor/messages etc down to 1 per frame, (adds all the movement together) mouse movements were huge slowdowns in code where uncessary really. Possibly want to do the same thing on the iphone, though it depends how many mouse moves we get in a frame...

Ortho cameras now have a variable range instead of being stuck at 100. Now we can make an ortho range 0-320, ie, 1 world unit = 1 pixel. Set in constructor, though we could change it on the fly for view-style scaling.

Made attach datum code in the render node a little more async. If we try to attach to a datum that is missing, it will keep trying on next draw. This helps attaching to nodes which aren't yet instanced and will be handy when we start to track datums.

Fixed some text-alignment code. (Still need a big text-test demo in the test game to make sure all the fonts are of the right size, alignment etc)
Align right, bottom, centre and middle are correct now.
Also changed the text alignment stuff to be 3D so you can use the Z of a datum (if it's a Box instead of Box2) for alignment. 2D still works, just aligned to z=0.