Moved linking of TootleLib-specific libs into their own libraries... ie. Box2D lib is now linked in the Physics lib instead of in the app, libpng is linked in the file sys lib.
Fixed templated LoadAsset() compile error.

Added some missing code to the load task so it would try again after a plain file failed to import. Added a missing check to see if an asset file needed importing, even if it was loaded before trying to load it into an asset.

Made WriteData in TBinary public again.

Added the ability to setup a TFile from plain TBinary data for a runtime file generated by binary data (eg. a stream) instead of from an asset or existing file.

Moved the TFileAsset Needs Import/Export flags out of the header (which is saved to the file, so they shouldn't be there) and into simple bools.

Added a check in the base TGame code so it won't call AddModes multiple times. It would only do that if a game has an asynchronous intialise.