Altered the social networking manager to be a templated manager allowing us to customise the manager per game for specific social networking platforms by simply altering the template parameter it uses.
This fixes an issue with building different projects using different social networking platforms where before we would have to alter a define in the engine each time we changed project to be able to test out a different type of social networking platform so all projects used the same one.

The only caveat for using this is that the global social networking manager pointer has to be defined gameside and cannot be used within any of the libraries atm but for this libraries case that isn't a problem and makes sense anyway as it would be project specific as to which type of system was used.

More work on this to come to remove a bunch of redundancy and add further routines for use in game, for unlocking acheivements etc.