Renamed GetData to GetAssetData in TFileAsset to help any ambiguity. Fixed some import/export flag settings in files when being copied and initialised, we should now do a little less unnecessary importing/exporting. Debug breaks will be thrown up when we're in awkward/unnecessary situations.
Along with Rocket changes:
Fixed the editor menus not coming up. This was a conflict to the changes to the storage of the widget data. The editor stored data with a ref of "node", which the widget also requires for the clickable node.
The scheme-editor node data is now "GrNode"(Graph node)