Added ability to prevent a scene node and/or it's physics or render nodes from being stored in the scheme.  This is mainly for nodes that are dynamically generated at runtime and aren't needed to be stored as part of the scheme, such as the exhaust and switch button child graphics.
To prevent a node from being stored simply attach an 'ignoredata' bool to the nodes binary data/properties and it will ignore storing the nodes and its children in a scheme when exporting.

NOTE: I've assumed that if you flag a node to be ignored in this way you would not want it's children to be stored as that would potentially break the graph when importing. Also atm it simply checks to see if this bit of data exists and ignores the node and children if so. At some stage we may need/want to check the bool being stored.