Renamed the TRef alphabet to be safe to send a plain tref string across with the web with url-encoding it. The symbols #? have been replaced with '! When creating a ref-string for an url/utf8 encoding, use GetUrlString from TRef which uses the alternative alphabet which replaces the space character with an asterisk*. The asterisk* can also be used in ref strings, and just turns into a space. This change also means we won't have trouble trying to create a file in windows with ? which is not an allowed symbol. (Just hadn't come across that problem yet I don't think)
Added GetDataHexString() to TBinary to get the data in a hex representation. I think this includes the array size (which the comments say it doesn't, will verify this)
Added WriteDataHexString() to TBinary which does the reverse and writes data into the TBinary from a string. Untested, and probably has problems as it doesn't specify explicitly if the data includes the array size data prefix...

Added some commented-out todo code in the IPodConnectionHttp.mm for me to use later when I add uploading to the ipod and fix an ipod crash.