Updated the Mac build configurations to enable Dead Code Stripping.
Fixed an issue with the render graph potentially adding multiple factories if the initialise of the render system lasts more than a frame.

Updated the initialise of a graph to only create a new root node if one doesn't exist - fixes an issue with allocating a new node (although no memory leak as it uses TPtr's) each frame of the render init (and any other graph init) if it lasts more than a frame.

Added the custom TootleOpenGLView to the render system that overloads the NSOpenGLView. This will be used for the main windows view.  Currently it only works when creating the view via the .nib file in the project. Eventually however this will be created via the code to allow us to customise it more easily without having to rely on a .nib file.  Also currently the window for a project is defined in the .nib file which will also be created in code eventually. WIP