Updated the app to create a window delegate (declared in the TootleRender lib) and assign it to the main window.  This is used to intercept window based events like closing the window.
Updated the creation of the memory pool on the Mac as it was triggering a EXC_BAD_ACCESS crash.  Previously it was all the time but now it seems to be every now and again so I haven;t fully corrected the issue but even in debug it's difficult to see what is causing the problem.  I'll investigate this further once the input handling is finished.

Added initial work on the input system using the Mac OS HID manager.  At the moment I have devices being enumerated and input data coming through as raw data from callbacks (traced out to the console) but this isn't being passed through to our input system as yet. WIP

Added a 'virtual' flag to the input device class.  This will be used to determine if a device has a physical counterpart or not.