Removed some redundant Mac specific routines for virtual device handling.
Implemented the enumeration of device objects, i.e. buttons and axis.  These are now found and setup correctly for a device attaching the required sensors that map to the device's object.  

TODO:  
Add HID queues for the device objects so that the object data can be accumulated between updates and then passed through the input system.