Sped up the array binary search a bit by providing a pointer to the data rather than having to use the virtual funcs every chop.
Added "char" is-binary-data-type declaration, the compiler treats char as differently to s8 (I presume becuase it's a macro and not compiler thing). Strings will now be duplicated/copied/imported/exported etc faster

Changed the graph RemoveChildren( FromNode ) function to use a node reference, this was to allow me to remove children from a node without having to find the TPtr (which was unnecessary anyway). This moves something from what we consider private to public, but it's a massive speed increase for something I was doing.

Added a break/check in TWidget to detect creating a widget with no rendertarget ref.