Fixed an issue with the screen manager not creating a screen instance for the Mac build.
Updated the iPod and Mac filesystem SetDirectory() and IsDirectoryValid() routines removing some redundant code and also ensuring the isdirectory flag returns true for the latter routine which was ignored before.

Added the following platform specific routines along with a generic TLFileSys wrapper for each:

Bool GetAssetDirectory(TTempString&);
Bool GetUserDirectory(TTempString&);

These are now used in the TApplication class to allow the removal of platform specific setups of the appropriate paths within the generic code and to simplify setups of additional projects without having to explicitly declare the app name for use in the paths.

NOTE: The platform macro checks will be removed after I add the equivalent for the PC as well.