Added a GetParentRef() routine to the base TLGraph::TGraphNode class.
Fixed an issue with the GetProperty() call for property access via the messaging not sending the response correctly.

Added some basic properties available for the TSceneNode_Object when requesting properties via the GetProperty messaging.

Added the ability for the scene nodes to instantiate child nodes when created via xml data using the <child> tag.  This mimics the render nodes but using the scene nodes and scenegraph.