Changed the use of NSURLRequest to NSMutableURLRequest.  This is becauses some of the methods called were not part of the NSURLRequest class so would likely do nothing or fail at runtime. I haven't tested this as I don;t know of an exampe that currently uses it but if you were having issues with this code then it'll be worth trying this again now to see if it now works.
NOTE: In objective-C if a class will not respond to a method call it will generate a wanring and not a compile error.  This is because the method call is more like a message to call the method in question so it is bound at runtime.  Therefore it is best to check all warning messages when altering objective-C code and do a full rebuild to double check them as they will often not be displayed again if the files are not recompiled.  

Any messages saying some class may not respond to some method means it is likely to not work as expected, so check the class definition to ensure the method exists and the parameters are correct.