Changed the memory allocation on the PC to zero the memory when building a debug version.  This helped find cases where memory was being allocated over memory that had previously been deallocated but still had pointers pointing to it.
Added the following callback hooks to the memory system when building in debug:

	FORCEINLINE void SetPreAllocCallback(memcallback pCallback);
	FORCEINLINE void SetPostAllocCallback(memcallback pCallback);
	FORCEINLINE void SetPreDeallocCallback(memcallback pCallback);
	FORCEINLINE void SetPostDeallocCallback(memcallback pCallback);

	FORCEINLINE void SetAllCallbacks(memcallback pCallback);

These allow you to test objects you find are being either overwritten or deleted prematurely more accurately especially when testing things like arrays where the arrays will rearrange themselves to be sorted when elements are added/removed which makes it quite difficult to spot what is actually causing the memory to change.  The last routine will simply set all callbacks to the passed in callback function.  It might be possible to change this to use the Publisher/Subscriber system instead of callbacks but I suspect this would seldom be used apart form specific cases so for now this system should suffice.

Added an integrity checking routine along with CHECK_ASSETARRAY_INTEGRITY define to be able to switch this on/off quickly.  This utilises the memory callback system and will test the asset managers list of assets for correctness - it checks the v-table pointer to see if it is valid using a few common values I found the v-table to end up being when it fails along with the loading state which has a set bounds within which it can be. Crude but works adequately for now and could possibly be added to the TPtrArray as a general dbeug validation routine at some stage.