Changed the RenderNode mesh and texture caches to use raw pointers instead of TPtr's.
This is because in some circumstances keeping a TPtr to an asset outside of the asset system would delete the asset on shutdown of the render node without the asset pointer in the asset system being removed correctly if the asset systems TPtr's had changed around whilst being sorted for instance.  

Additionally in most cases the use of the caches was via the raw pointer anyway so it makes more sense to simply be a raw pointer in addition to the implication this is for speed purposes so removing the TPtr indirection will improve it's use slightly.  

The render node's GetMeshAsset() and GetTextureAsset() now return a raw pointer inline with this change and I've also changed all overloads of GetMeshAsset() to match.

The timeline instance class also stored a TPtr to assets which could potentially have deleted the asset prematurely so I've changed that as well to simply be a raw pointer.

NOTE: For both the render node and timeline instance, if they allow for the asset ref to change at runtime we will likely need a private OnAssetRefChange() type routine to handle updating the 'cache' pointer and eventually when assets gets unloaded during gameplay we will need to handle the asset being removed fromt he asset system by intercepting the file system change notifications.  I would add this to TRAC but it currently seems to be offline :(