Updated the debug TTY output to use a TTempString as a buffer rather than immediately outputting to the console.  This improves the performances of debug builds slightly by reducing the frequency of sending the debug text to the TTY until the buffer is full.
When a debug break occurs this will flush the buffer and immediately print the message so critical imformation can still be seen in these cases.

Added new Mac and IPod platform specific debug files that the debug Print() and Break() have been moved into from the MacCore.mm and IPodCore.mm.  PC version to follow shortly.  This eases making future changes to the TTY output handling.

Removed the debug break from the memory system destructor - now that the buffer is used, which is a static variable, the buffer may or may not exist when TLDebug_Break() or TLDebug_Print() is called via other static variable destructors, so can potentially fail.  This is probably the only likely static variable that currently calls these routines during destruction so I've removed it for now pending the addition of some better memory tracking.