Fixed small meory leak in the LocalFileSys::SetDirectory(TString&) routine on both the Mac and iPod.
Re-wrote the Mac LocalFileSys::LoadFileSys(const char*) routine to only do a shallow search of the specified directory for the filesystem.  Previously this would search all subdirectories as well which is not required and caused problems with files trying to be loaded unnecessarily or incorrectly.  This doesn't occur  on the iPod version because all the files are stored flat within the root of the .app package.