Implemented the memory pattern filling on allocation.  This will fill newly allocated memory with the TLMemory::Debug_AllocPattern defined as 0xfa during a debug build only.  This is now cross platform (not checked on Mac or iPod yet) and should aid with spotting uninitialised variables.
NOTE: I've also added the code to fill memory with a TLMemory::Debug_DeallocPattern on deallocation but whilst testing this I noticed that the size being used is incorrect - currently always giving the size of the void* i.e. just the size of a pointer and not the size of the memory pointed to by the pointer.  There doesn't appear to be an easy way to determine the size at this late stage of deallocation and I can't find out how the free() method manages to figure it out so for now it will simply fill the pattern for the size of a pointer only.  Additionally this means the memory used tracking is currently incorrectly updated.