Fixed the determination of the memory size being incorrect for deallocation of memory but only on the PC.  This will query the heap with the void* to determine the size that will be removed. Now correct the memory used tracking is now reliable on the PC and still highlights we have some memory going astray within the system on shut down which will be looked into over time.
In addition I have implemented the tracing out of the memory allocations still active in the heap on the PC.  This will go through and print to the console (in debug) what memory is still in use and the memory address.  This should be useful in finding some of the memory leaks we still have when shutting down.