Fixed an issue with the TGraph class not actually releasing the nodes in the graph correctly on shutdown - used the new heap debug output to spot this problem.
Previously this would simply NULL the root node which would leave all of the child nodes referencing each other as a 'lost' tree of nodes and they would not then get released correctly when the graph destructed.  Now the root node is explicitly removed from the graph on shutdown which also removes all child nodes correctly.  Additionally the graph request queue is also explicitly emptied on shutdown which prevents some debug errors about nodes being in the queue when shutting down the tree.