Big change to the TArray class to use two template policy classes for sorting and array data allocation.  These should simplify a few things and allow for a great deal of flexibility in the future especially for testing things like different sort methods.  I intend to break this class (and it's derivatives) down further by adding two more policies for searching and storage that hopefully will remove the need for a lot of the virtual functions that provide the equivalent functionality and improve performance/efficiency as well.
Anyway, everything should work as before and performance should be comparable or better  as the routines should inline well and I have also removed a few unnecessary redundant checks like for the sort function pointer that would only be needed for sorted arrays that are now setup at compile time only for the arrays that use a sorting policy class.  However, just in case, I have left all of the old functionality in the classes and added USE_SORT_POLICY and USE_ALLOCATOR_POLICY defines to allow them to be turned on/off in case a problem arises whilst I continue to test the games.  I will remove these and any redundant code, such as the sort function pointer, from the TArray class itself in about a week or so when I am happy that I haven't inadvertently broken anything.  The TPtrArray, TFixedArray and TKeyArray have all been updated to allow them to be customised where possible using these new policy classes.

Deprecated the TArray ElementRand routines.  This is so that the TArray class isn't dependant on the TLMaths random functions requiring the TLMaths.h file to always be included.  Instead I have added a TArrayMisc.h file with two equivalent template routines for getting a random element that can be used in there place.   

Added a templated TPointerArray class that is defined as TArray<TYPE*> to allow for potential optimisations of this type of array across the board in the future, such as using an SOA for the pointer allocation.  I've changed the TPublisher and TSubscriber classes to use the TPointerArray class and I will replace other arrays of TArray<TYPE*> as I come across them.

Removed the TPublisher SendMessageReverse which was rarely used and didn't have the desired effect.

Added the ability to change the order of an event channels publication which effectively re-arranges the array of subscribers.  This is now used to allow the shutdown message event channel to send it's message in the correct order.

Abstracted the TLMaths::TTransform out of the TLMaths.h file.  This was due to this class using the TArray class that caused cyclical header include issues.  Where appropriate the new TTransform.h file include has been added.

Removed the TMemory.h file dependancy of TCoreMisc.h and place the PointerToInteger() and IntegerToPointer() routines in the TLMemory.cpp file instead.

Moved the TLMaths_ShapeRef_### defines into a TLMathsMisc.h file.  This is due to other header changes that meant the TRef wasn't defined when including TLMaths.h so rather than always include TRef.h I have moved them and updated any files as necessary that use these defines.